+++ docs/source/changelog.rst
@@ -10,6 +10,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Added
 ^^^^^
 * Clarification on common status issues in documentation
+* Check if output.log exists at runtime
 
 Fixed
 ^^^^^^